+++ Backend-Stuff/gamifiedcalendar/gamifiedcalendar/src/main/java/com/swe/gamifiedcalendar/AccountRepository.java
@@ -2,6 +2,6 @@ package com.swe.gamifiedcalendar;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AccountRepository extends CrudRepository<Account, Integer>{
-    Account findById(int accountId);
+public interface AccountRepository extends CrudRepository<Account, Long>{
+    Account findById(long accountId);
 }
